requirements.txt
Django>=5.1,<6.0
djangorestframework>=3.16,<4.0
djangorestframework-simplejwt>=5.4,<6.0
django-cors-headers>=4.7,<5.0
python-dotenv>=1.0.1

# AI & ML
openai-whisper>=20240930
google-genai

# YouTube
yt-dlp>=2024.12

# Utilities
Pillow>=11.0

# Testing & Code Quality
coverage>=7.6
flake8>=7.0
